requirements.txt
Flask
gensim
pandas
spacy
scispacy
kaggle
elasticsearch_dsl>=7.0.0